refactor(#4172): introduce IDataExplorerWidgetStorage and migrate getDataExplorerWidgetStorage to typed interface
