+++ streampipes-rest/src/main/java/org/apache/streampipes/rest/impl/datalake/KioskDashboardDataLakeResource.java
@@ -28,8 +28,8 @@ import org.apache.streampipes.model.datalake.SpQueryResult;
 import org.apache.streampipes.model.datalake.param.ProvidedRestQueryParams;
 import org.apache.streampipes.model.monitoring.SpLogMessage;
 import org.apache.streampipes.rest.core.base.impl.AbstractAuthGuardedRestResource;
-import org.apache.streampipes.storage.api.CRUDStorage;
 import org.apache.streampipes.storage.api.IDataExplorerDashboardStorage;
+import org.apache.streampipes.storage.api.IDataExplorerWidgetStorage;
 import org.apache.streampipes.storage.api.IPermissionStorage;
 import org.apache.streampipes.storage.management.StorageDispatcher;
 
@@ -53,7 +53,7 @@ public class KioskDashboardDataLakeResource extends AbstractAuthGuardedRestResou
   private final IDataExplorerSchemaManagement dataExplorerSchemaManagement;
   private final IDataExplorerDashboardStorage dashboardStorage =
       StorageDispatcher.INSTANCE.getNoSqlStore().getDataExplorerDashboardStorage();
-  private final CRUDStorage<DataExplorerWidgetModel> dataExplorerWidgetStorage;
+  private final IDataExplorerWidgetStorage dataExplorerWidgetStorage;
   private final IPermissionStorage permissionStorage;
 
   public KioskDashboardDataLakeResource() {
